resthits/requirements.txt
Flask==1.1.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
Flask-Testing==0.7.1
marshmallow==3.3.0
psycopg2-binary==2.8.4
pytest==5.3.1
pytest-black==0.3.7
pytest-isort==0.3.1
pytest-flakes==4.0.0
python-slugify==2.0.1
factory_boy==2.12.0
Werkzeug==0.16.1
